% Generated by roxygen2 (4.0.1): do not edit by hand
\name{actionPaint}
\alias{actionPaint}
\title{actionPaint}
\usage{
actionPaint(rawimg, segmimg)
}
\arguments{
\item{rawimg}{An Image object with the raw frame data}

\item{segmimg}{An Image object with the segmented objects}
}
\value{
An Image object that combines raw and segmented images, with objects painted singularly with different colours
}
\description{
actionPaint
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

