% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{addParticles}
\alias{addParticles}
\title{Combines the information from a raw \code{Frames} object and the corresponding preprocessed one}
\usage{
addParticles(raw.frames, binary.frames, col = NULL)
}
\arguments{
\item{raw.frames}{A \code{Frames} object containing the raw images}

\item{binary.frames}{A \code{Frames} object with the preprocessed versions of the images (e.g. segmented)}

\item{col}{A color character string, to select which color will be used for drawing the contours of the particles. If not specified, it will default according to the objects provided}
}
\value{
A \code{Frames} object, whose images are the combination of the raw images with the segmented objects drawn on them
}
\description{
All objects are painted with a unique colour - for sake of speed
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

