% Generated by roxygen2 (4.0.1): do not edit by hand
\name{axesInfo}
\alias{axesInfo}
\title{axesInfo}
\usage{
axesInfo(framelist)
}
\arguments{
\item{framelist}{A FrameList object}
}
\value{
A list object, containing the extremes of the field of interest (x-y-z, where z is time)
}
\description{
Auxiliary function to return the dimensions of the field of interest
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

