% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{candidate.platelets}
\alias{candidate.platelets}
\title{candidate.platelets}
\description{
A sample ParticleList object
}
\details{
The sample ParticleList object is constituted by the platelets identified from the MesenteriumSubset data
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

