% Generated by roxygen2 (4.0.2): do not edit by hand
\name{computeMSD}
\alias{computeMSD}
\title{computeMSD}
\usage{
computeMSD(sx, sy, until = 4)
}
\arguments{
\item{sx}{x axis positions along the trajectory}

\item{sy}{y axis positions along the trajectory}

\item{until}{how many points should be included in the Mean Squared Displacement curve}
}
\value{
A numeric vector containing the values of the MSD
}
\description{
Calculates the Mean Squared Displacement for a trajectory
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

