% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createChannelsFrameList}
\alias{createChannelsFrameList}
\title{createChannelsFrameList}
\usage{
createChannelsFrameList(framelist)
}
\arguments{
\item{framelist}{A FrameList object}
}
\value{
A ChannelsFrameList object, which is a list of 3 FrameList objects, named respectively red, green and blue
}
\description{
Creates a ChannelsFrameList object from a FrameList object, decomposed in the acquired channels
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

