% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cut.FrameList}
\alias{cut.FrameList}
\title{cut.FrameList}
\usage{
\method{cut}{FrameList}(x, cutLeft = 5, cutRight = 5, cutUp = 5,
  cutDown = 5, cutAll = 0, testing = FALSE, ...)
}
\arguments{
\item{x}{An input FrameList object}

\item{cutLeft}{Amount of pixels to be cut at the side}

\item{cutRight}{Amount of pixels to be cut at the side}

\item{cutUp}{Amount of pixels to be cut at the side}

\item{cutDown}{Amount of pixels to be cut at the side}

\item{cutAll}{Amount of pixels to be cut at all sides. Overrides the single side values}

\item{testing}{Logical, whether to just test the cropping or to actually perform it. Default set to FALSE}

\item{...}{Arguments to be passed to methods}
}
\value{
A FrameList object, with cropped frames in the image slot
}
\description{
Performs cropping on the FrameList object, selecting how many pixels should be cut on each side
}
\details{
Cropping can be performed with careful choice of all cutting sides, or cropping a single value from
all sides
}
\examples{
load(file.path(system.file("extra", package="flowcatchR"),"MesenteriumSubset.RData"))
cut(MesenteriumSubset)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

