% Generated by roxygen2 (4.0.1): do not edit by hand
\name{display.TrajectoryList}
\alias{display.TrajectoryList}
\title{display.TrajectoryList}
\usage{
\method{display}{TrajectoryList}(trajectorylist, framelist)
}
\arguments{
\item{trajectorylist}{A TrajectoryList object}

\item{framelist}{A FrameList object, used here to identify the limits of the region of interest}
}
\description{
Provides a visual representation of a TrajectoryList object
}
\details{
Based on the rgl library, the function extracts the region of interests from the dimensions of an image of the FrameList object,
and afterwards plots the x-y-time representation of the identified trajectories
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

