% Generated by roxygen2 (4.0.1): do not edit by hand
\name{evaluateTrajectoryList}
\alias{evaluateTrajectoryList}
\title{evaluateTrajectoryList}
\usage{
evaluateTrajectoryList(trajectorylist, rawframelist, preprocessedframelist)
}
\arguments{
\item{trajectorylist}{A TrajectoryList object}

\item{rawframelist}{A FrameList object with raw images}

\item{preprocessedframelist}{A FrameList object with preprocessed frames}
}
\value{
A new TrajectoryList object
}
\description{
A pseudo-interactive function to inspect and accept/reject trajectories of a TrajectoryList object
}
\details{
The user gets the trajectories displayed singularly, and for each a keyboard input is required.
The slot corresponding to the user feedback is interactively filled in. The TrajectoryList object is then ready for further operations
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

