% Generated by roxygen2 (4.0.2): do not edit by hand
\name{export.particles}
\alias{export.particles}
\title{Exports a \code{ParticleList} object}
\usage{
export.particles(particlelist, dir = tempdir(),
  nameStub = "testExport_particles_")
}
\arguments{
\item{particlelist}{A \code{ParticleList} object}

\item{dir}{The path of the folder where the particle lists should be written}

\item{nameStub}{The stub for the file name, that will be used as a prefix for the exported particle lists}
}
\value{
Particle list files are written in the desired location
}
\description{
Writes the particles contained in the \code{particles} data frame slot of the \code{ParticleList} object elements.
A track of the provenience of the particles is stored as a comment line above the header
}
\examples{
data("candidate.platelets")
\dontrun{export.particles(candidate.platelets)}
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

