% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractParticles}
\alias{extractParticles}
\title{extractParticles}
\usage{
extractParticles(framelistRaw, framelistPreprocessed = NULL, channel = "",
  areaThresholdMin = 5, areaThresholdMax = 100)
}
\arguments{
\item{framelistRaw}{A FrameList object with the raw images (mandatory)}

\item{framelistPreprocessed}{A FrameList object with preprocessed images (optional, if not provided gets produced with standard default parameters)}

\item{channel}{Character string. The channel to perform the operations on. Can be "red", "green" or "blue"}

\item{areaThresholdMin}{Size in pixels of the minimum area needed to detect the object as a potential particle of interest}

\item{areaThresholdMax}{Size in pixels of the maximum area allowed to detect the object as a potential particle of interest}
}
\value{
A ParticleList object, containing all detected particles for each frame
}
\description{
Extract particles from the images of a FrameList object.
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

