% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filterParticles}
\alias{filterParticles}
\title{filterParticles}
\usage{
filterParticles(particlelist, areaThresholdMin = 1, areaThresholdMax = 1000)
}
\arguments{
\item{particlelist}{A ParticleList object. A LinkedParticleList object can also be provided as input, yet the returned object will be a ParticleList object that}

\item{areaThresholdMin}{Size in pixels of the minimum area needed to detect the object as a potential particle of interest}

\item{areaThresholdMax}{Size in pixels of the maximum area allowed to detect the object as a potential particle of interest}
}
\value{
A ParticleList object
}
\description{
Performs filtering on a ParticleList object
}
\details{
According to parameters of interests, such as size, eccentricity/shape, filters out the particles that do not
satisfy the indicated requirements
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

