% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flowcatchR.R
\docType{package}
\name{flowcatchR}
\alias{flowcatchR}
\alias{flowcatchR-package}
\title{flowcatchR: analyzing time-lapse microscopy imaging, from detection to tracking}
\description{
A toolset to analyze in vivo microscopy imaging data focused on tracking
flowing blood cells.
}
\details{
flowcatchR is a set of tools to analyze in vivo microscopy imaging
data, focused on tracking flowing blood cells. It guides the steps from
segmentation to calculation of features, filtering out particles not of
interest, providing also a set of utilities to help checking the quality of
the performed operations (e.g. how good the segmentation was). The main
novel contribution investigates the issue of tracking flowing cells such as
in blood vessels, to categorize the particles in flowing, rolling and
adherent. This classification is applied in the study of phenomena such as
hemostasis and study of thrombosis development.
}
\author{
Federico Marini \email{marinif@uni-mainz.de},
Johanna Mazur \email{mazur@uni-mainz.de},
Harald Binder \email{binderh@uni-mainz.de}, 2015

Maintainer: Federico Marini \email{marinif@uni-mainz.de}
}

