% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{frameList1}
\alias{frameList1}
\title{frameList1}
\description{
A sample FrameList object provided
}
\details{
The sample FrameList object is constituted by a subset of a time-lapse intravital microscopy imaging dataset.
Green channel marks leukocytes, red channel focuses on blood platelets.
Images are kindly provided by Sven Jaeckel (\email{Sven.Jaeckel@unimedizin-mainz.de}).
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

