% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Frames-class}
\alias{Frames-class}
\title{Frames class}
\description{
S4 class for storing information on multiple images belonging to the same time-lapse experiment.
It is designed as a subclass of the existing Image class from the EBImage package
}
\section{Slots}{

\describe{
\item{\code{channel}}{A character vector, can be 'red','green','blue' or 'all' (if in color mode)}
}}

