% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{Frames}
\alias{Frames}
\title{Constructor for a \code{Frames} object}
\usage{
Frames(x, channel)
}
\arguments{
\item{x}{A multi-dimensional \code{Image} object}

\item{channel}{A character vector, can be 'red','green','blue' or 'all' (if in color mode)}
}
\value{
The created \code{Frames} object.
}
\description{
Constructor for a \code{Frames} object
}
\examples{
data("MesenteriumSubset")
inputImg <- Image(MesenteriumSubset)
Frames(inputImg,"red")
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

