% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fullInspection.FrameList}
\alias{fullInspection.FrameList}
\title{fullInspection.FrameList}
\usage{
fullInspection.FrameList(framelist, nframes = 10)
}
\arguments{
\item{framelist}{A FrameList object}

\item{nframes}{The number of frames to display}
}
\description{
Display entirely a FrameList object
}
\details{
All frames of a FrameList are displayed in the browser, and are interactively navigable.
Default number of shown frames is 10, but in this case there is no upper limit to display.
Care should be taken with large FrameList objects
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

