% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate.TrajectoryList}
\alias{generate.TrajectoryList}
\title{generate.TrajectoryList}
\usage{
generate.TrajectoryList(particlelist, provideExtraDetails = FALSE, ...)
}
\arguments{
\item{particlelist}{A ParticleList object}

\item{provideExtraDetails}{Logical, currently not used - could be introduced for providing additional info on the trajectories}

\item{...}{Arguments to be passed to methods}
}
\value{
A TrajectoryList object
}
\description{
Generates a TrajectoryList object from a LinkedParticleList
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

