% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initialize.LinkedParticleSet}
\alias{initialize.LinkedParticleSet}
\title{Initialize a \code{ParticleList} object for subsequent linking/tracking}
\usage{
initialize.LinkedParticleSet(particleset, linkrange = 1)
}
\arguments{
\item{linkrange}{The number of frames to look for candidate particles potentially belonging to the same track}

\item{particlelist}{A ParticleList object}
}
\value{
A ParticleList object with slots dedicated for the tracking pre-filled
}
\description{
Initialize a \code{ParticleList} object for subsequent linking/tracking
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

