% Generated by roxygen2 (4.0.1): do not edit by hand
\name{initialize.ParticleList}
\alias{initialize.ParticleList}
\title{initialize.ParticleList}
\usage{
initialize.ParticleList(particlelist, linkrange = 1)
}
\arguments{
\item{particlelist}{A ParticleList object}

\item{linkrange}{The number of frames to look for candidate particles potentially belonging to the same track}
}
\value{
A ParticleList object with slots dedicated for the tracking pre-filled
}
\description{
Initialize a ParticleList object for subsequent linking/tracking
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

