% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inspect.FrameList}
\alias{inspect.FrameList}
\title{inspect.FrameList}
\usage{
inspect.FrameList(framelist, nframes = 4)
}
\arguments{
\item{framelist}{A FrameList object}

\item{nframes}{The number of frames to display}
}
\description{
Explore the first frames of a FrameList
}
\details{
The first frames of a FrameList are displayed in the browser, and are interactively navigable.
Default number of shown frames is 4, can be set maximum to 8, as this function is purely for a
first inspection
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

