% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inspect.frames}
\alias{inspect.frames}
\title{inspect.frames}
\usage{
inspect.frames(framelist, nframes = NULL, inspectAll = FALSE,
  display.method = "browser")
}
\arguments{
\item{framelist}{A FrameList object}

\item{nframes}{The number of frames to display (default value: 6)}

\item{inspectAll}{Logical, whether to inspect all frames (overriding the default of 10 that can be used also when inspectAll is FALSE)}

\item{display.method}{Method for displaying, can be either "raster" or "browser". Defaults to browser, by opening a window in the browser}
}
\description{
Explore the first frames of a FrameList
}
\details{
The first frames of a FrameList are displayed in the browser, and are interactively navigable.
Default number of shown frames is 4, can be set maximum to 8, as this function is purely for a
first inspection
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

