% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inspect.trajectories}
\alias{inspect.trajectories}
\title{inspect.trajectories}
\usage{
inspect.trajectories(trajectorylist, rawframelist, preprocessedframelist,
  trajID, browse = TRUE)
}
\arguments{
\item{trajectorylist}{A TrajectoryList object}

\item{rawframelist}{A FrameList object with raw images}

\item{preprocessedframelist}{A FrameList object with preprocessed frames}

\item{trajID}{A numeric value, storing the ID of trajectory which is under inspection}

\item{browse}{Logical, whether to automatically open the painted trajectory to inspect}
}
\value{
A new TrajectoryList object -NO! Returns now a FrameList object, and this can be used to evaluate the trajectory
}
\description{
A function to inspect single trajectories from a TrajectoryList object
}
\details{
Basically it add the contours to just one object, and returns a FrameList object which is a subset of the input one
The length of the returned object corresponds exactly the length of the identified trajectory, so uninteresting frames will be skipped

OLD: A pseudo-interactive function to inspect and accept/reject trajectories of a TrajectoryList object
OLD: The user gets the trajectories displayed singularly, and for each a keyboard input is required.
The slot corresponding to the user feedback is interactively filled in. The TrajectoryList object is then ready for further operations
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

