% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{KinematicsFeatures-class}
\alias{KinematicsFeatures-class}
\title{KinematicsFeatures class}
\description{
S4 class for storing information on all kinematics features identified for a single trajectory
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A list storing the information for the kinematics features}
}}

