% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{KinematicsFeaturesSet-class}
\alias{KinematicsFeaturesSet-class}
\title{KinematicsFeaturesSet}
\description{
S4 class for storing information on all kinematics features identified for all trajectories. Single
\code{KinematicsFeatures} objects are the element of the main list
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A list storing the information for the sets of kinematics features}
}}

