% Generated by roxygen2 (4.0.2): do not edit by hand
\name{length.Frames}
\alias{length.Frames}
\title{Creates a \code{ChannelsFrameList} object from a \code{FrameList} object, decomposed in the acquired channels}
\usage{
\method{length}{Frames}(x)
}
\arguments{
\item{framelist}{A \code{FrameList} object}
}
\value{
A \code{ChannelsFrameList} object, which is a list of 3 \code{FrameList} objects, named respectively \code{red}, \code{green} and \code{blue}
}
\description{
Creates a \code{ChannelsFrameList} object from a \code{FrameList} object, decomposed in the acquired channels
}
\examples{
data("MesenteriumSubset")
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
export
}

