% Generated by roxygen2 (4.0.1): do not edit by hand
\name{link.ParticleList}
\alias{link.ParticleList}
\title{link.ParticleList}
\usage{
link.ParticleList(particlelist, L, R = 2, epsilon1 = 0.1, epsilon2 = 2,
  lambda1 = 1, lambda2 = 1, penaltyFunction = penaltyFunctionGenerator(),
  nframes, verboseOutput = F, useIntensityChange = TRUE,
  useAreaChange = TRUE)
}
\arguments{
\item{particlelist}{A ParticleList object}

\item{L}{Maximum number of pixels an object can move in two consecutive frames}

\item{R}{Linkrange, i.e. the number of consecutive frames to search for potential candidate links}

\item{epsilon1}{A numeric value, to be used in the formula. Jitter for allowing angular displacements}

\item{epsilon2}{A numeric value, to be used in the formula. Jitter for allowing spatial displacements}

\item{lambda1}{A numeric value. Multiplicative factor for the penalty function}

\item{lambda2}{A numeric value. Multiplicative factor applied to the angular displacement}

\item{penaltyFunction}{A function structured in such a way to be applied as penalty function in the linking}

\item{nframes}{Numeric value. Number of frames where the linking should be performed}

\item{verboseOutput}{Logical, whether the output should report additional intermediate steps. For debugging use mainly}

\item{useIntensityChange}{Logical, whether to include also intensity change of the particles in the cost function calculation}

\item{useAreaChange}{Logical, whether to include also area change of the particles in the cost function calculation}
}
\value{
A LinkedParticleList object
}
\description{
Performs linking of the particles by tracking them through the frames
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}
\references{
Sbalzarini et al.?
}

