% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{MesenteriumSubset}
\alias{MesenteriumSubset}
\title{MesenteriumSubset}
\description{
A sample FrameList object provided
}
\details{
The sample FrameList object is constituted by a subset of a time-lapse intravital microscopy imaging dataset.
Green channel marks leukocytes, red channel focuses on blood platelets. 20 frames are provided in this subset.
Images are kindly provided by Sven J\"ackel (\email{Sven.J\"ackel@unimedizin-mainz.de}).
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

