% Generated by roxygen2 (4.0.1): do not edit by hand
\name{newFrameList}
\alias{newFrameList}
\title{newFrameList
constructor for a FrameList object}
\usage{
newFrameList(nframes, imgsLocation)
}
\arguments{
\item{nframes}{Number of frames that will constitute the FrameList object}

\item{imgsLocation}{Vector of strings containing the locations where the (raw) images are to be found}
}
\description{
newFrameList
constructor for a FrameList object
}

