% Generated by roxygen2 (4.0.1): do not edit by hand
\name{paintTrajectory}
\alias{paintTrajectory}
\title{paintTrajectory}
\usage{
paintTrajectory(trajectorylist, rawframelist, preprocessedframelist, trajId)
}
\arguments{
\item{trajectorylist}{A TrajectoryList object}

\item{rawframelist}{A FrameList object with raw images}

\item{preprocessedframelist}{A FrameList object with preprocessed frames}

\item{trajId}{Numeric value, the ID of the trajectory}
}
\value{
A new FrameList object
}
\description{
Creates a FrameList objects containing raw information, combined with the segmented images and the relative trajectory under analysis
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

