% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ParticleSet-class}
\alias{ParticleSet-class}
\title{ParticleSet class}
\description{
S4 class for storing information on particles detected in distinct frames.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A list storing the information for the particles}

\item{\code{channel}}{A character vector, can be 'red','green', or 'blue'. It refers to which channel the particles were detected}
}}

