% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tracking.R
\name{penaltyFunctionGenerator}
\alias{penaltyFunctionGenerator}
\title{Generate a penalty function}
\usage{
penaltyFunctionGenerator(epsilon1 = 0.1, epsilon2 = 2, lambda1 = 1,
  lambda2 = 1)
}
\arguments{
\item{epsilon1}{A numeric value, to be used in the formula. Jitter for allowing angular displacements}

\item{epsilon2}{A numeric value, to be used in the formula. Jitter for allowing spatial displacements}

\item{lambda1}{A numeric value. Multiplicative factor for the penalty function}

\item{lambda2}{A numeric value. Multiplicative factor applied to the angular displacement}
}
\value{
A function object, to be used as penalty function
}
\description{
A function to generate penalty functions to use while linking particles
}
\examples{
custom.function <- penaltyFunctionGenerator(epsilon1=0.1,epsilon2=6,lambda1=1.5,lambda2=0)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

