% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot2D.TrajectoryList}
\alias{plot2D.TrajectoryList}
\title{plot2D.TrajectoryList}
\usage{
plot2D.TrajectoryList(x, framelist, trajIDs = NULL, ...)
}
\arguments{
\item{x}{A TrajectoryList object}

\item{framelist}{A FrameList object, used here to identify the limits of the region of interest}

\item{trajIDs}{A vector containing the ids of the desired trajectories}

\item{...}{Arguments to be passed to methods}
}
\description{
Provides a bird's eye view of a TrajectoryList object on a bidimensional space
}
\details{
Independent from the rgl library, the function extracts the region of interests from the dimensions of an image of the FrameList object,
and afterwards plots the x-y-time representation of the identified trajectories on a 2d plane. It is possible to subset the TrajectoryList
object with the IDs of the desired trajectories
}
\examples{
load(file.path(system.file("extra", package="flowcatchR"),"MesenteriumSubset.RData"))
load(file.path(system.file("extra", package="flowcatchR"),"candidate.platelets.RData"))
platelets.trajectories <- trajectories(candidate.platelets)
plot2D.TrajectoryList(platelets.trajectories,MesenteriumSubset)
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

