% Generated by roxygen2 (4.0.1): do not edit by hand
\name{preprocess.FrameList}
\alias{preprocess.FrameList}
\title{preprocess.FrameList}
\usage{
\method{preprocess}{FrameList}(x, brushSize = 3, brushShape = "disc",
  adaptOffset = 0.15, adaptWinWidth = 10, adaptWinHeight = 10,
  kernSize = 3, kernShape = "disc", watershedTolerance = 1,
  watershedRadius = 1, displayprocessing = FALSE, areaThresholdMin = 5,
  areaThresholdMax = 100, ...)
}
\arguments{
\item{x}{A FrameList object}

\item{brushSize}{Size in pixels of the brush to be used for initial smoothing}

\item{brushShape}{Shape of the brush to be used for initial smoothing}

\item{adaptOffset}{Offset to be used in the adaptive thresholding step}

\item{adaptWinWidth}{Width of the window for the adaptive thresholding step}

\item{adaptWinHeight}{Height of the window for the adaptive thresholding step}

\item{kernSize}{Size in pixels of the kernel used for morphological operations}

\item{kernShape}{Shape of the kernel used for morphological operations}

\item{watershedTolerance}{Tolerance allowed in performing the watershed-based segmentation}

\item{watershedRadius}{Radius for the watershed-based segmentation}

\item{displayprocessing}{Logical, whether to display intermediate steps while performing preprocessing. Dismissed currently, it could increase runtime a lot}

\item{areaThresholdMin}{Size in pixels of the minimum area needed to detect the object as a potential particle of interest}

\item{areaThresholdMax}{Size in pixels of the maximum area allowed to detect the object as a potential particle of interest}

\item{...}{Arguments to be passed to methods}
}
\value{
A FrameList object, whose frame images are the preprocessed versions of the input images
}
\description{
Preprocessing function for FrameList objects
}
\details{
FrameList objects are processed according to the chosen set of parameters. Many of them refer directly to
existing EBImage functions, please see the corresponding help for additional information
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

