% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\usage{
preprocess(x, ...)
}
\arguments{
\item{x}{A FrameList or a ChannelsFrameList object}

\item{...}{Arguments to be passed to methods, such as channel and/or preprocessing parameters}
}
\description{
Generic preprocessing function
}
\details{
Can be applied to FrameList or ChannelsFrameList objects. ChannelsFrameList are then subset to the chosen channel,
and the method preprocess.FrameList is then applied, with its set of parameters
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

