% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.FrameList}
\alias{print.FrameList}
\title{Display conveniently a \code{FrameList} object}
\usage{
\method{print}{FrameList}(x, ...)
}
\arguments{
\item{x}{A \code{FrameList} object}

\item{...}{Arguments to be passed to methods}
}
\description{
Display conveniently a \code{FrameList} object
}
\examples{
data("MesenteriumSubset")
print(MesenteriumSubset)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

