% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.LinkedParticleList}
\alias{print.LinkedParticleList}
\title{print.LinkedParticleList}
\usage{
\method{print}{LinkedParticleList}(x, ...)
}
\arguments{
\item{x}{A LinkedParticleList object}

\item{...}{Arguments to be passed to methods}
}
\description{
Method for displaying conveniently a LinkedParticleList object
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

