% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.TrajectoryList}
\alias{print.TrajectoryList}
\title{print.TrajectoryList}
\usage{
\method{print}{TrajectoryList}(x, ...)
}
\arguments{
\item{x}{A TrajectoryList object}

\item{...}{Arguments to be passed to methods}
}
\description{
Method for displaying conveniently a TrajectoryList object
}
\examples{
load(file.path(system.file("extra", package="flowcatchR"),"candidate.platelets.RData"))
platelets.trajectories <- trajectories(candidate.platelets)
print(platelets.trajectories)
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

