% Generated by roxygen2 (4.0.2): do not edit by hand
\name{repmat}
\alias{repmat}
\title{repmat}
\usage{
repmat(a, n, m)
}
\arguments{
\item{a}{The matrix to copy}

\item{n}{The n value for the tiling}

\item{m}{The m value for the tiling}
}
\value{
Creates a large matrix consisting of an m-by-n tiling of copies of a.
}
\description{
Function equivalent for MATLAB's repmat - Replicate and tile arrays
}
\details{
A more flexible and stylish alternative to replicate the behaviour of the repmat function of MATLAB
}
\author{
Robin Hankin, 2001
}
\references{
http://cran.r-project.org/doc/contrib/R-and-octave.txt
}

