% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reproducibleColorLabels}
\alias{reproducibleColorLabels}
\title{reproducibleColorLabels}
\usage{
reproducibleColorLabels(x, normalize = TRUE)
}
\arguments{
\item{x}{an Image object in Grayscale color mode or an array containing object masks}

\item{normalize}{Logical, if TRUE normalizes the resulting color image}
}
\value{
An Image object with color coded objects
}
\description{
Color codes the labels of object masks by a random permutation
}
\details{
The labels of object masks are coloured this time in a "random but reproducible way", by setting a seed at the beginning of the function
}
\examples{
\dontrun{
oneImg <- subset.FrameList(MesenteriumSubset,framesToKeep = 1)
showMe(preprocess(oneImg)[[1]]$image)
showMe(reproducibleColorLabels(preprocess(oneImg)[[1]]$image[,,1]))
}
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}
\references{
colorLabels in the EBImage package
}

