% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{shinyFlow}
\alias{shinyFlow}
\title{Shiny application for exploring the features and parameters provided by \code{flowcatchR}}
\usage{
shinyFlow()
}
\value{
The Shiny Application is launched in the web browser
}
\description{
Launches a Shiny Web Application for interactive data exploration. Default data loaded are
the frames from the \code{MesenteriumSubset} object, custom values can be inserted by typing
the location of the data stored in a local folder. The Application is structured in a variety
of tabs that mirror the steps in the usual workflow in time-lapse microscopy images. These can
allow the user to interactively explore the parameters and their effect in the reactive
framework provided by Shiny.
}
\examples{
\dontrun{shinyFlow()}
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2015
}
