% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,KinematicsFeatures-method}
\alias{show,KinematicsFeatures-method}
\title{Displaying conveniently a \code{KinematicsFeatures} object}
\usage{
\S4method{show}{KinematicsFeatures}(object)
}
\arguments{
\item{object}{A \code{KinematicsFeatures} object}

\item{...}{Arguments to be passed to methods}
}
\value{
This returns an invisible \code{NULL}.
}
\description{
Displaying conveniently a \code{KinematicsFeatures} object
}
\examples{
data("candidate.platelets")
platelets.trajectories <- trajectories(candidate.platelets)
traj11features <- kinematics(platelets.trajectories,trajectoryIDs = 11)
print(traj11features)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

