% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,KinematicsFeaturesSet-method}
\alias{show,KinematicsFeaturesSet-method}
\title{Display conveniently a \code{KinematicsFeatureSetList} object}
\usage{
\S4method{show}{KinematicsFeaturesSet}(object)
}
\arguments{
\item{object}{A \code{KinematicsFeatureSetList} object}

\item{...}{Arguments to be passed to methods}
}
\description{
Display conveniently a \code{KinematicsFeatureSetList} object
}
\examples{
data("candidate.platelets")
platelets.trajectories <- trajectories(candidate.platelets)
alltrajs.features <- kinematics(platelets.trajectories)
print(alltrajs.features)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

