% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\name{show,LinkedParticleSet-method}
\alias{show,LinkedParticleSet-method}
\title{Display conveniently a \code{LinkedParticleSet} object}
\usage{
\S4method{show}{LinkedParticleSet}(object)
}
\arguments{
\item{object}{A \code{LinkedParticleSet} object}
}
\value{
This returns an invisible \code{NULL}.
}
\description{
Display conveniently a \code{LinkedParticleSet} object
}
\examples{
data("candidate.platelets")
linked.platelets <- link.particles(candidate.platelets,L=26,R=3,epsilon1=0,
epsilon2=0,lambda1=1,lambda2=0,penaltyFunction=penaltyFunctionGenerator(),
include.area=FALSE)
print(linked.platelets)

}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}
