% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,ParticleSet-method}
\alias{show,ParticleSet-method}
\title{Display conveniently a \code{ParticleList} object}
\usage{
\S4method{show}{ParticleSet}(object)
}
\arguments{
\item{x}{A \code{ParticleList} object}

\item{...}{Arguments to be passed to methods}
}
\description{
Display conveniently a \code{ParticleList} object
}
\examples{
data("candidate.platelets")
print(candidate.platelets)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

