% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show,TrajectorySet-method}
\alias{show,TrajectorySet-method}
\title{Display conveniently a \code{TrajectoryList} object}
\usage{
\S4method{show}{TrajectorySet}(object)
}
\arguments{
\item{x}{A \code{TrajectoryList} object}

\item{...}{Arguments to be passed to methods}
}
\description{
Display conveniently a \code{TrajectoryList} object
}
\examples{
data("candidate.platelets")
platelets.trajectories <- trajectories(candidate.platelets)
print(platelets.trajectories)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

