% Generated by roxygen2 (4.0.1): do not edit by hand
\name{showMe}
\alias{showMe}
\title{Displays with raster}
\usage{
showMe(imgObject, dispMet = "raster", ...)
}
\arguments{
\item{imgObject}{an \code{Image} object}

\item{dispMet}{Set to default as "raster", could be "browser"}

\item{...}{Arguments to be passed to display}
}
\description{
Displays images with R native functions, makes direct use of EBImage's display
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

