% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toCartesianCoords}
\alias{toCartesianCoords}
\title{toCartesianCoords}
\usage{
toCartesianCoords(Theta, Radius)
}
\arguments{
\item{Theta}{The Theta angle}

\item{Radius}{The radius value in polar coordinates}
}
\value{
A list containing Theta and Radius, as in polar coordinates
}
\description{
Converts polar coordinates to cartesian coordinates
}
\details{
Conversion from (radius,theta) to (x,y)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

