% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trajectories}
\alias{trajectories}
\title{Generate trajectories}
\usage{
trajectories(particlelist, provideExtraDetails = FALSE, ...)

trajectories(particlelist, provideExtraDetails = FALSE, ...)
}
\arguments{
\item{particlelist}{A (\code{Linked})\code{ParticleList} object}

\item{provideExtraDetails}{Logical, currently not used - could be introduced for providing additional info on the trajectories}

\item{...}{Arguments to be passed to methods}

\item{particleset}{A (\code{Linked})\code{ParticleList} object}

\item{verbose}{Logical, currently not used - could be introduced for providing additional info on the trajectories}

\item{...}{Arguments to be passed to methods}
}
\value{
A \code{TrajectoryList} object

A \code{TrajectoryList} object
}
\description{
Generates a \code{TrajectoryList} object from a (\code{Linked})\code{ParticleList}

Generates a \code{TrajectoryList} object from a (\code{Linked})\code{ParticleList}
}
\examples{
data("candidate.platelets")
platelets.trajectories <- trajectories(candidate.platelets)
data("candidate.platelets")
platelets.trajectories <- trajectories(candidate.platelets)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014

Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

