% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TrajectorySet-class}
\alias{TrajectorySet-class}
\title{TrajectorySet class}
\description{
S4 class for storing information on the trajectories identified, including whether there were gaps, the number of points, and more
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A list storing the information for the particles}

\item{\code{channel}}{A character vector, can be 'red','green', or 'blue'. It refers to which channel the particles were detected}
}}

