% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add.contours}
\alias{add.contours}
\title{add.contours}
\usage{
add.contours(raw.frames, binary.frames, trajectories, trajId)
}
\arguments{
\item{raw.frames}{A FrameList object with raw images}

\item{binary.frames}{A FrameList object with preprocessed frames}

\item{trajectories}{A trajectories object}

\item{trajId}{Numeric value, the ID of the trajectory}
}
\value{
A new FrameList object
}
\description{
Creates a FrameList objects containing raw information, combined with the segmented images and the relative trajectory under analysis
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

