% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trajectories.R
\name{axesInfo}
\alias{axesInfo}
\title{Info on the dimensions of the FOV}
\usage{
axesInfo(frames)
}
\arguments{
\item{frames}{A \code{Frames} object}
}
\value{
A list object, containing the extremes of the field of interest (x-y-z, where z is time)
}
\description{
Auxiliary function to return the dimensions of the field of interest
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

