% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{candidate.platelets}
\alias{candidate.platelets}
\title{A sample \code{ParticleList} object}
\description{
The sample \code{ParticleList} object is constituted by the platelets identified from the \code{MesenteriumSubset} data
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

