% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowcatchR.R
\docType{data}
\name{candidate.platelets}
\alias{candidate.platelets}
\title{A sample \code{ParticleSet} object}
\description{
The sample \code{ParticleSet} object is constituted by the platelets identified from the \code{MesenteriumSubset} data
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

