% Generated by roxygen2 (4.0.2): do not edit by hand
\name{channel.Frames}
\alias{channel.Frames}
\title{Channel extraction for \code{objects}}
\usage{
channel.Frames(frames, mode)
}
\arguments{
\item{frames}{A \code{Frames} object}

\item{mode}{A character value specifying the target mode for conversion.}
}
\description{
\code{channel}
}

