% Generated by roxygen2 (4.0.1): do not edit by hand
\name{combine.preprocessedFrameList}
\alias{combine.preprocessedFrameList}
\title{combine.preprocessedFrameList}
\usage{
combine.preprocessedFrameList(rawframelist, preprocessedframelist, col = NULL)
}
\arguments{
\item{rawframelist}{A FrameList object containing the raw images}

\item{preprocessedframelist}{A FrameList object with the preprocessed versions of the images (e.g. segmented)}

\item{col}{A color character string, to select which color will be used for drawing the contours of the particles. If not specified, it will default according to the objects provided}
}
\value{
A FrameList object, whose images are the combination of the raw images with the segmented objects drawn on them
}
\description{
Combines the information from a raw FrameList object and the corresponding preprocessed one
}
\details{
All objects are painted with a unique colour - for sake of speed
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

