% Generated by roxygen2 (4.0.1): do not edit by hand
\name{combineWcolor.preprocessedFrameList}
\alias{combineWcolor.preprocessedFrameList}
\title{combineWcolor.preprocessedFrameList}
\usage{
combineWcolor.preprocessedFrameList(rawframelist, preprocessedframelist)
}
\arguments{
\item{rawframelist}{A FrameList object containing the raw images}

\item{preprocessedframelist}{A FrameList object with the preprocessed versions of the images (e.g. segmented)}
}
\value{
A FrameList object, whose images are the combination of the raw images with the segmented objects drawn on them,
painted singularly with different colours
}
\description{
Combines the information from a raw FrameList object and the corresponding preprocessed one,
but this time every object is painted with a different colour
}
\details{
Every object is now shown with a different colour. Care should be taken, as this function is rather slower
than combine.preprocessedFrameList
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

