% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crop.Frames}
\alias{crop.Frames}
\title{#Combines raw and segmented images
#
#}
\usage{
crop.Frames(frames, cutLeft = 5, cutRight = 5, cutUp = 5, cutDown = 5,
  cutAll = 0, testing = FALSE, ...)
}
\arguments{
\item{cutLeft}{Amount of pixels to be cut at the side}

\item{cutRight}{Amount of pixels to be cut at the side}

\item{cutUp}{Amount of pixels to be cut at the side}

\item{cutDown}{Amount of pixels to be cut at the side}

\item{cutAll}{Amount of pixels to be cut at all sides. Overrides the single side values}

\item{testing}{Logical, whether to just test the cropping or to actually perform it. Default set to \code{FALSE}}

\item{...}{Arguments to be passed to methods}

\item{rawimg}{An \code{Image} object with the raw frame data
#}

\item{segmimg}{An \code{Image} object with the segmented objects
#
#}

\item{rawframelist}{A \code{FrameList} object containing the raw images}

\item{preprocessedframelist}{A \code{FrameList} object with the preprocessed versions of the images (e.g. segmented)}

\item{x}{An input \code{FrameList} object}
}
\value{
An \code{Image} object that combines raw and segmented images, with objects painted singularly with different colours
#
#

A \code{FrameList} object, whose images are the combination of the raw images with the segmented objects drawn on them,
painted singularly with different colours

A \code{FrameList} object, with cropped frames in the \code{image} slot
}
\description{
#Combines raw and segmented images
#
#
}
\examples{
data("MesenteriumSubset2")
crop.Frames(MesenteriumSubset2)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
Combines the information from a raw \code{FrameList} object and the corresponding preprocessed one,
but this time every object is painted with a different colour

Every object is now shown with a different colour. Care should be taken, as this function is rather slower
than combine.preprocessedFrameList

Federico Marini, \email{marinif@uni-mainz.de}, 2014
Cut borders of a \code{FrameList} object

Performs cropping on the \code{FrameList} object, selecting how many pixels should be cut on each side

Cropping can be performed with careful choice of all cutting sides, or cropping a single value from
all sides

Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

