% Generated by roxygen2 (4.0.1): do not edit by hand
\name{export.frames}
\alias{export.frames}
\title{export.frames}
\usage{
export.frames(framelist, folder = "/Users/fede/TEMP/exportFrameList/",
  nameStub = "testExport", createGif = FALSE,
  removeAfterCreatingGif = TRUE)
}
\arguments{
\item{framelist}{A FrameList object}

\item{folder}{The path of the folder where the image should be written}

\item{nameStub}{The stub for the file name, that will be used as a prefix for the exported images}

\item{createGif}{Logical, whether to create or not an animated .gif file}

\item{removeAfterCreatingGif}{Logical, whether to remove the single exported .png images after creating the single .gif}
}
\value{
Image files are written in the desired location
}
\description{
Exports a FrameList object
}
\details{
Writes the images contained in the image slot of the FrameList object elements.
The images can be exported as single frames, or as a .gif image that is composed
by the single frames.
}
\examples{
load(file.path(system.file("extra", package="flowcatchR"),"MesenteriumSubset.RData"))
\dontrun{export.frames(MesenteriumSubset,nameStub="subset_export_",createGif=TRUE,removeAfterCreatingGif=FALSE)}
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

