% Generated by roxygen2 (4.0.1): do not edit by hand
\name{export.particles}
\alias{export.particles}
\title{export.particles}
\usage{
export.particles(particlelist, folder = "",
  nameStub = "testExport_particles_")
}
\arguments{
\item{particlelist}{A ParticleList object}

\item{folder}{The path of the folder where the particle lists should be written}

\item{nameStub}{The stub for the file name, that will be used as a prefix for the exported particle lists}
}
\value{
Particle list files are written in the desired location
}
\description{
Exports a ParticleList object
}
\details{
Writes the particles contained in the particles data frame slot of the ParticleList object elements.
A track of the provenience of the particles is stored as a comment line above the header
}
\examples{
load(file.path(system.file("extra", package="flowcatchR"),"candidate.platelets.RData"))
\dontrun{export.particles(candidate.platelets)}
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

