% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractKinematics.traj}
\alias{extractKinematics.traj}
\title{extractKinematics.traj}
\usage{
extractKinematics.traj(trajectorylist, trajectoryID,
  acquisitionFrequency = 30, scala = 50)
}
\arguments{
\item{trajectorylist}{A TrajectoryList object}

\item{trajectoryID}{The ID of a single trajectory}

\item{acquisitionFrequency}{The frame rate of acquisition for the images, in
milliseconds}

\item{scala}{The value of micro(?)meters to which each single pixel
  corresponds}
}
\value{
A KinematicsFeatureSet object
}
\description{
Calculate a set of kinematics parameter from a single trajectory
}
\details{
The computed set of parameters include delta.x, delta.t and delta.v
(displacements and instantaneous velocity), totalTime, totalDistance,
distStartToEnd, curvilinearVelocity, straightLineVelocity and
linearityForwardProgression, Mean Squared Displacement and velocity
autocorrelation
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

