% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tracking.R
\name{initialize.LinkedParticleSet}
\alias{initialize.LinkedParticleSet}
\title{Initialize a \code{ParticleSet} object for subsequent linking/tracking}
\usage{
initialize.LinkedParticleSet(particleset, linkrange = 1)
}
\arguments{
\item{particleset}{A \code{ParticleSet} object}

\item{linkrange}{The number of frames to look for candidate particles potentially belonging to the same track}
}
\value{
A \code{ParticleSet} object with slots dedicated for the tracking pre-filled
}
\description{
Initialize a \code{ParticleSet} object for subsequent linking/tracking
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

