% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initialize.ParticleList}
\alias{initialize.ParticleList}
\title{Initialize a \code{ParticleList} object for subsequent linking/tracking}
\usage{
initialize.ParticleList(particlelist, linkrange = 1)
}
\arguments{
\item{particlelist}{A ParticleList object}

\item{linkrange}{The number of frames to look for candidate particles potentially belonging to the same track}
}
\value{
A ParticleList object with slots dedicated for the tracking pre-filled
}
\description{
Initialize a \code{ParticleList} object for subsequent linking/tracking
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

