% Generated by roxygen2 (4.0.2): do not edit by hand
\name{inspect.Frames}
\alias{inspect.Frames}
\title{Explore the first frames of a \code{FrameList}}
\usage{
inspect.Frames(frames, nframes = NULL, display.method = "browser",
  verbose = FALSE)
}
\arguments{
\item{nframes}{The number of frames to display (default value: 6)}

\item{display.method}{Method for displaying, can be either \code{raster} or \code{browser}. Defaults to \code{browser}, by opening a window in the browser}

\item{verbose}{Logical, whether to provide additional output on the command line alongside with the images themselves}

\item{framelist}{A \code{FrameList} object}

\item{inspectAll}{Logical, whether to inspect all frames (overriding the default of 10 that can be used also when \code{inspectAll} is FALSE)}
}
\description{
The first frames of a \code{FrameList} are displayed in the browser, and are interactively navigable.
Default number of shown frames is 4, can be set maximum to 8, as this function is purely for a
first inspection
}
\examples{
data("MesenteriumSubset")
\dontrun{inspect.frames(MesenteriumSubset)}
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

