% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{inspect.Frames}
\alias{inspect.Frames}
\title{Explore the frames of a \code{Frames}}
\usage{
inspect.Frames(frames, nframes = NULL, display.method = "browser",
  verbose = FALSE)
}
\arguments{
\item{frames}{A \code{Frames} object}

\item{nframes}{The number of frames to display (default value: \code{NULL}, all are displayed )}

\item{display.method}{Method for displaying, can be either \code{raster} or \code{browser}. Defaults to \code{browser}, by opening a window in the browser}

\item{verbose}{Logical, whether to provide additional output on the command line alongside with the images themselves}
}
\value{
\code{inspect.Frames} returns an invisible \code{NULL}.
}
\description{
The first frames of a \code{Frames} are displayed in the browser, and are interactively navigable.
}
\examples{
data("MesenteriumSubset")
\dontrun{inspect.Frames(MesenteriumSubset)}
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

