% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{length.Frames}
\alias{length.Frames}
\title{Compute the length of render frames in a \code{Frames} object}
\usage{
\method{length}{Frames}(x)
}
\arguments{
\item{x}{A \code{Frames} object}
}
\value{
An integer number
}
\description{
Compute the length of render frames in a \code{Frames} object
}
\examples{
data("MesenteriumSubset")
length(MesenteriumSubset)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

