% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{LinkedParticleSet-class}
\alias{LinkedParticleSet-class}
\title{LinkedParticleSet class}
\description{
S4 class for storing information of particles after they have been tracked. It inherits the slots
from the \code{ParticleSet} class.
}
\section{Slots}{

\describe{
\item{\code{tracking}}{A list storing all necessary information for the tracking algorithm to work, and for providing the information
to the function to determine the trajectories}
}}

