% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowcatchR.R
\docType{data}
\name{MesenteriumSubset}
\alias{MesenteriumSubset}
\title{A sample \code{Frames} object}
\description{
The sample \code{Frames} object is constituted by a subset of a time-lapse intravital microscopy imaging dataset.
Green channel marks leukocytes, red channel focuses on blood platelets. 20 frames are provided in this subset.
Images are kindly provided by Sven Jaeckel (\email{Sven.Jaeckel@unimedizin-mainz.de}).
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

