% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preprocess.ChannelsFrameList}
\alias{preprocess.ChannelsFrameList}
\title{preprocess.ChannelsFrameList}
\usage{
\method{preprocess}{ChannelsFrameList}(x, channel = "", ...)
}
\arguments{
\item{x}{A ChannelsFrameList object}

\item{channel}{Character string. The channel to perform the operations on. Can be "red", "green" or "blue"}

\item{...}{Arguments to be passed to methods}
}
\value{
A FrameList object, whose frame images are the preprocessed versions of the input images
}
\description{
Preprocessing function for ChannelsFrameList objects
}
\details{
ChannelsFrameList are then subset to the chosen channel, and the method preprocess.FrameList is then applied, with its set of parameters
}
\examples{
data("MesenteriumSubset")
preprocess(channels(MesenteriumSubset),channel = "red")
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

