% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preprocess.Frames}
\alias{preprocess.Frames}
\title{Preprocessing images}
\usage{
preprocess.Frames(frames, brush.size = 3, brush.shape = "disc",
  at.offset = 0.15, at.wwidth = 10, at.wheight = 10, kern.size = 3,
  kern.shape = "disc", ws.tolerance = 1, ws.radius = 1,
  displayprocessing = FALSE, ...)
}
\arguments{
\item{brush.size}{Size in pixels of the brush to be used for initial smoothing}

\item{brush.shape}{Shape of the brush to be used for initial smoothing}

\item{at.offset}{Offset to be used in the adaptive thresholding step}

\item{at.wwidth}{Width of the window for the adaptive thresholding step}

\item{at.wheight}{Height of the window for the adaptive thresholding step}

\item{kern.size}{Size in pixels of the kernel used for morphological operations}

\item{kern.shape}{Shape of the kernel used for morphological operations}

\item{ws.tolerance}{Tolerance allowed in performing the watershed-based segmentation}

\item{ws.radius}{Radius for the watershed-based segmentation}

\item{displayprocessing}{Logical, whether to display intermediate steps while performing preprocessing. Dismissed currently, it could increase runtime a lot}

\item{...}{Arguments to be passed to methods, such as channel and/or preprocessing parameters}

\item{x}{A \code{FrameList} or a \code{ChannelsFrameList} object}

\item{x}{A \code{ChannelsFrameList} object}

\item{channel}{Character string. The channel to perform the operations on. Can be \code{red}, \code{green} or \code{blue}}

\item{...}{Arguments to be passed to methods}

\item{x}{A \code{FrameList} object}

\item{...}{Arguments to be passed to methods}
}
\value{
A \code{FrameList} object, whose frame images are the preprocessed versions of the input images

A \code{FrameList} object, whose frame images are the preprocessed versions of the input images
}
\description{
Can be applied to \code{FrameList} or \code{ChannelsFrameList} objects. \code{ChannelsFrameList} are then subset to the chosen channel,
and the method \code{preprocess.FrameList} is then applied, with its set of parameters
}
\examples{
data("MesenteriumSubset")
preprocess(channels(MesenteriumSubset),channel = "red")
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
#' Preprocessing function for \code{ChannelsFrameList} objects

\code{ChannelsFrameList} are then subset to the chosen channel, and the method \code{preprocess.FrameList} is then applied, with its set of parameters

Federico Marini, \email{marinif@uni-mainz.de}, 2014
Preprocessing function for \code{FrameList} objects

\code{FrameList} objects are processed according to the chosen set of parameters. Many of them refer directly to
existing \code{EBImage} functions, please see the corresponding help for additional information

Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

