% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.FrameList}
\alias{print.FrameList}
\title{print.FrameList}
\usage{
\method{print}{FrameList}(x, ...)
}
\arguments{
\item{x}{A FrameList object}

\item{...}{Arguments to be passed to methods}
}
\description{
Method for displaying conveniently a FrameList object
}
\examples{
load(file.path(system.file("extra", package = "flowcatchR"), "MesenteriumSubset.RData"))
print(MesenteriumSubset)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

