% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.KinematicsFeatureSet}
\alias{print.KinematicsFeatureSet}
\title{Displaying conveniently a \code{KinematicsFeatureSet} object}
\usage{
\method{print}{KinematicsFeatureSet}(x, ...)
}
\arguments{
\item{x}{A \code{KinematicsFeatureSet} object}

\item{...}{Arguments to be passed to methods}
}
\description{
Displaying conveniently a \code{KinematicsFeatureSet} object
}
\examples{
data("candidate.platelets")
platelets.trajectories <- trajectories(candidate.platelets)
traj11features <- kinematics(platelets.trajectories,trajectoryID = 11)
print(traj11features)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

