% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.KinematicsFeatureSet}
\alias{print.KinematicsFeatureSet}
\title{print.KinematicsFeatureSet}
\usage{
\method{print}{KinematicsFeatureSet}(x, ...)
}
\arguments{
\item{x}{A KinematicsFeatureSet object}

\item{...}{Arguments to be passed to methods}
}
\description{
Method for displaying conveniently a KinematicsFeatureSet object
}
\examples{
load(file.path(system.file("extra", package="flowcatchR"),"candidate.platelets.RData"))
platelets.trajectories <- trajectories(candidate.platelets)
traj11features <- kinematics(platelets.trajectories,trajectoryID = 11)
print(traj11features)
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

