% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.KinematicsFeatureSetList}
\alias{print.KinematicsFeatureSetList}
\title{print.KinematicsFeatureSetList}
\usage{
\method{print}{KinematicsFeatureSetList}(x, ...)
}
\arguments{
\item{x}{A KinematicsFeatureSetList object}

\item{...}{Arguments to be passed to methods}
}
\description{
Method for displaying conveniently a KinematicsFeatureSetList object
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

