% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.ParticleList}
\alias{print.ParticleList}
\title{print.ParticleList}
\usage{
\method{print}{ParticleList}(x, ...)
}
\arguments{
\item{x}{A ParticleList object}

\item{...}{Arguments to be passed to methods}
}
\description{
Method for displaying conveniently a ParticleList object
}
\examples{
load(file.path(system.file("extra", package="flowcatchR"),"candidate.platelets.RData"))
print(candidate.platelets)
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

