% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.ParticleList}
\alias{print.ParticleList}
\title{print.ParticleList}
\usage{
\method{print}{ParticleList}(x, ...)
}
\arguments{
\item{x}{A ParticleList object}

\item{...}{Arguments to be passed to methods}
}
\description{
Method for displaying conveniently a ParticleList object
}
\examples{
data("candidate.platelets")
print(candidate.platelets)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

