% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.frames}
\alias{read.frames}
\title{read.frames}
\usage{
read.frames(imgsLocation, nframes)
}
\arguments{
\item{imgsLocation}{Vector of strings containing the locations where the (raw) images are to be found}

\item{nframes}{Number of frames that will constitute the FrameList object}
}
\value{
An object of the \code{FrameList} class, which holds the info on a list of frames, specifying for each the following elements:
\item{image}{The \code{Image} object containing the image itself}
\item{location}{The complete path to the location of the original image}
}
\description{
constructor for a FrameList object
}
\details{
This function is used to create a FrameList object from a vector of images.
The number of frames is also specified, as just a subset of the images can be used for this
}

