% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{rotate.Frames}
\alias{rotate.Frames}
\title{Rotates all images in a \code{Frames} object}
\usage{
rotate.Frames(frames, angle, testing = FALSE,
  output.origin = c(dim(frames)[1]/3, dim(frames)[2]/3),
  output.dim = c(dim(frames)[1] * 1.5, dim(frames)[2] * 1.5))
}
\arguments{
\item{frames}{A \code{Frames} object}

\item{angle}{The rotation angle (clockwise) specified in degrees}

\item{testing}{Logical, whether to just test the rotation or to actually perform it. Default set to \code{FALSE}}

\item{output.origin}{A vector of 2 numbers indicating the dimension of the output image, as in the rotate function}

\item{output.dim}{A vector of 2 numbers indicating the output coordinates of the origin in pixels, as in the \code{rotate} function}
}
\value{
A \code{Frames} object containing the rotated frames
}
\description{
Rotation is performed exploiting the rotate function of the \code{EBImage} package. Could be automated if support for coordinate/pixel interaction is included
}
\examples{
data("MesenteriumSubset")
rotate.Frames(MesenteriumSubset,angle = 40)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

