% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select.Frames}
\alias{select.Frames}
\title{Extracts subsets of frames from a \code{FrameList} object}
\usage{
select.Frames(frames, framesToKeep = 1, ...)
}
\arguments{
\item{framesToKeep}{A vector containing the indexes of the frames to keep in the selection}

\item{...}{Arguments to be passed to methods}

\item{x}{A \code{FrameList} object}
}
\value{
A \code{FrameList} object, composed by the subset of frames of the input \code{FrameList}
}
\description{
An input \code{FrameList} object is subject to subsetting. This function is useful e.g. when the trajectory of interest
is presenting gaps (i.e. does not actually include a frame)
}
\examples{
data("MesenteriumSubset")
subset(MesenteriumSubset, framesToKeep = c(1:10, 14:20))
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

