% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,Frames-method}
\alias{show,Frames-method}
\title{Display conveniently a \code{Frames} object}
\usage{
\S4method{show}{Frames}(object)
}
\arguments{
\item{object}{A \code{Frames} object}

\item{...}{Arguments to be passed to methods}
}
\value{
This returns an invisible \code{NULL}.
}
\description{
Display conveniently a \code{Frames} object
}
\examples{
data("MesenteriumSubset")
print(MesenteriumSubset)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

