% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,KinematicsFeaturesSet-method}
\alias{show,KinematicsFeaturesSet-method}
\title{Display conveniently a \code{KinematicsFeatureSet} object}
\usage{
\S4method{show}{KinematicsFeaturesSet}(object)
}
\arguments{
\item{object}{A \code{KinematicsFeatureSet} object}

\item{...}{Arguments to be passed to methods}
}
\value{
This returns an invisible \code{NULL}.
}
\description{
Display conveniently a \code{KinematicsFeatureSet} object
}
\examples{
data("candidate.platelets")
platelets.trajectories <- trajectories(candidate.platelets)
alltrajs.features <- kinematics(platelets.trajectories)
print(alltrajs.features)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

