% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,ParticleSet-method}
\alias{show,ParticleSet-method}
\title{Display conveniently a \code{ParticleSet} object}
\usage{
\S4method{show}{ParticleSet}(object)
}
\arguments{
\item{object}{A \code{ParticleSet} object}

\item{...}{Arguments to be passed to methods}
}
\value{
This returns an invisible \code{NULL}.
}
\description{
Display conveniently a \code{ParticleSet} object
}
\examples{
data("candidate.platelets")
print(candidate.platelets)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

