% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,TrajectorySet-method}
\alias{show,TrajectorySet-method}
\title{Display conveniently a \code{TrajectorySet} object}
\usage{
\S4method{show}{TrajectorySet}(object)
}
\arguments{
\item{object}{A \code{TrajectorySet} object}

\item{...}{Arguments to be passed to methods}
}
\value{
This returns an invisible \code{NULL}.
}
\description{
Display conveniently a \code{TrajectorySet} object
}
\examples{
data("candidate.platelets")
platelets.trajectories <- trajectories(candidate.platelets)
print(platelets.trajectories)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

