% Generated by roxygen2 (4.0.2): do not edit by hand
\name{subset.FrameList}
\alias{subset.FrameList}
\title{subset.FrameList}
\usage{
\method{subset}{FrameList}(x, framesToKeep, ...)
}
\arguments{
\item{x}{A FrameList object}

\item{framesToKeep}{A vector containing the indexes of the frames to keep in the selection}

\item{...}{Arguments to be passed to methods}
}
\value{
A FrameList object, composed by the subset of frames of the input FrameList
}
\description{
Extracts a subset of frames from a FrameList object
}
\details{
An input FrameList object is subject to subsetting. This function is useful e.g. when the trajectory of interest
is presenting gaps (i.e. does not actually include a frame)
}
\examples{
load(file.path(system.file("extra", package = "flowcatchR"), "MesenteriumSubset.RData"))
subset(MesenteriumSubset, framesToKeep = c(1:10, 14:20))
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

