% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kinematics.R
\name{toCartesianCoords}
\alias{toCartesianCoords}
\title{Converts polar coordinates to cartesian coordinates}
\usage{
toCartesianCoords(Theta, Radius)
}
\arguments{
\item{Theta}{The Theta angle}

\item{Radius}{The radius value in polar coordinates}
}
\value{
A list containing Theta and Radius, as in polar coordinates
}
\description{
Conversion from (radius,theta) to (x,y)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

