% Generated by roxygen2 (4.0.1): do not edit by hand
\name{toPolarCoords}
\alias{toPolarCoords}
\title{toPolarCoords}
\usage{
toPolarCoords(x, y)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}
}
\value{
A list containing Theta and Radius, as in polar coordinates
}
\description{
Converts cartesian coordinates to polar coordinates
}
\details{
Conversion from (x,y) to (radius,theta)
}
\author{
Federico Marini, \email{federico.marini@uni-mainz.de}, 2014
}

