% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kinematics.R
\name{toPolarCoords}
\alias{toPolarCoords}
\title{Converts cartesian coordinates to polar coordinates}
\usage{
toPolarCoords(x, y)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}
}
\value{
A list containing Theta and Radius, as in polar coordinates
}
\description{
Conversion from (x,y) to (radius,theta)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

