% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trajectories}
\alias{trajectories}
\title{trajectories}
\usage{
trajectories(particlelist, provideExtraDetails = FALSE, ...)
}
\arguments{
\item{particlelist}{A (Linked)ParticleList object}

\item{provideExtraDetails}{Logical, currently not used - could be introduced for providing additional info on the trajectories}

\item{...}{Arguments to be passed to methods}
}
\value{
A TrajectoryList object
}
\description{
Generates a TrajectoryList object from a (Linked)ParticleList
}
\examples{
load(file.path(system.file("extra", package="flowcatchR"),"candidate.platelets.RData"))
platelets.trajectories <- trajectories(candidate.platelets)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

