% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trajectories.R
\name{trajectories}
\alias{trajectories}
\title{Generate trajectories}
\usage{
trajectories(particleset, verbose = FALSE, ...)
}
\arguments{
\item{particleset}{A (\code{Linked})\code{ParticleSet} object}

\item{verbose}{Logical, currently not used - could be introduced for providing additional info on the trajectories}

\item{...}{Arguments to be passed to methods}
}
\value{
A \code{TrajectorySet} object
}
\description{
Generates a \code{TrajectorySet} object from a (\code{Linked})\code{ParticleSet}
}
\examples{
data("candidate.platelets")
platelets.trajectories <- trajectories(candidate.platelets)
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2014
}

